Expose android compatible popup API to common
