+++ compose/ui/ui/src/commonMain/kotlin/androidx/compose/ui/window/SecureFlagPolicy.kt
@@ -16,25 +16,24 @@
 
 package androidx.compose.ui.window
 
-import android.view.WindowManager
-
 /**
- * Policy on setting [WindowManager.LayoutParams.FLAG_SECURE] on a window.
+ * Policy on setting [android.view.WindowManager.LayoutParams.FLAG_SECURE] on a window.
  */
 enum class SecureFlagPolicy {
     /**
-     * Inherit [WindowManager.LayoutParams.FLAG_SECURE] from the parent window and pass it on the
+     * Inherit [android.view.WindowManager.LayoutParams.FLAG_SECURE] from the parent window and pass it on the
      * window that is using this policy.
      */
     Inherit,
 
     /**
-     * Forces [WindowManager.LayoutParams.FLAG_SECURE] to be set on the window that is using this
+     * Forces [android.view.WindowManager.LayoutParams.FLAG_SECURE] to be set on the window that is using this
      * policy.
      */
     SecureOn,
+
     /**
-     * No [WindowManager.LayoutParams.FLAG_SECURE] will be set on the window that is using this
+     * No [android.view.WindowManager.LayoutParams.FLAG_SECURE] will be set on the window that is using this
      * policy.
      */
     SecureOff
